refactor login-page component
